Add function save to take and take_answer table

+++ src/main/java/com/group10/contestPlatform/dtos/quiz/usersubmitanswer/UserSubmitAnswerRequest.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserSubmitAnswerRequest {
-    private int quizQuestion;
-    private int selectedAnswer;
+    private long quizQuestion;
+    private long selectedAnswer;
 }
